+++ SDLGroup4/SDLGroup4/Animator.h
@@ -14,7 +14,7 @@ namespace Engine
 
 		void Trigger(const std::string& name);
 		void Stop();
-		void DisplayAnimation(int x, int y);//(Vector2D position);
+		void DisplayAnimation(float x, float y);//(Vector2D position);
 
 		std::vector<Animation*> Animations;
 		Animation* currenAnimation = nullptr;
